NewsStatus/requirements.txt
absl-py==0.8.1
aiowebsocket==1.0.0.dev2
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.8.0
backports.functools-lru-cache==1.5
billiard==3.3.0.23
cachetools==3.1.1
celery==3.1.25
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cma==2.7.0
configparser==4.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.1.4
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
Django==1.11.26
django-celery==3.2.1
django-crontab==0.7.1
enum34==1.1.6
Flask==1.1.1
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gevent==1.4.0
google-auth==1.6.3
google-auth-oauthlib==0.4.1
graphviz==0.13
greenlet==0.4.15
grpcio==1.24.3
gunicorn==19.7.1
hyperlink==19.0.0
idna==2.6
imageio==2.4.1
importlib-metadata==0.23
incremental==17.5.0
ipaddress==1.0.17
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.3
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
kombu==3.0.37
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
more-itertools==5.0.0
moviepy==1.0.1
MySQL-python==1.2.5
nltk==3.4
numpy==1.16.4
oauthlib==3.1.0
objgraph==3.4.1
opencv-python==4.1.1.26
paddlehub==1.2.1
paddlepaddle==1.6.0
pandas==0.24.2
parsel==1.5.2
pathlib2==2.3.5
Pillow==6.2.1
prettytable==0.7.2
proglog==0.1.9
Protego==0.1.15
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycrypto==2.6.1
pycryptodome==3.9.0
PyDispatcher==2.0.5
pygobject==3.26.1
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
queuelib==1.5.0
rarfile==3.1
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scandir==1.10.0
scipy==1.2.1
Scrapy==1.8.0
SecretStorage==2.3.1
service-identity==18.1.0
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.5.4
tb-nightly==2.1.0a20191102
tb-paddle==0.3.6
tqdm==4.37.0
Twisted==19.7.0
urllib3==1.25.6
vine==1.3.0
visualdl==1.3.0
w3lib==1.21.0
websocket-client==0.56.0
Werkzeug==0.16.0
xlrd==1.2.0
zipp==0.6.0
zope.interface==4.6.0
